Resolve wp_^(page_menu nav_menu) conflict. Replace sidebar "Pages" section with more general "Navigation".
Doesn't make sense to double the page menu, and would need to conditionalize the whole nav_menu <li> to detect if a menu was set or not. Instead, take advantage of nav_menu's fallback:  IF user sets a custom navigation menu, show it. IF NOT, it falls back to page_menu, giving something close to the "Pages" list it replaces.

This changes a translated string, the page/nav menu title.