Initial support and use of post thumbnails.
Gimmicky use of the feature for now - only on
search results. Codex/Theme_Review lists
thumbnail feature as required as of wp2.9.

If it exists, thumb img is displayed left of its post's
result excerpt in search.php search results template.

entry-title needed a clear:both to overcome the float
if a post has thumb img but no content text.